requirements.txt
aiohttp
telethon
telethon-tgcrypto
python-dotenv==0.18.0
